+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -8,6 +8,7 @@ using Libplanet.Blockchain;
 using Libplanet.Blockchain.Policies;
 using Libplanet.Blocks;
 using Libplanet.Crypto;
+using Libplanet.Store;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tests.Store;
 using Libplanet.Tx;
@@ -725,6 +726,80 @@ namespace Libplanet.Tests.Blockchain
                 callCount <= 1,
                 $"GetBlocksStates() was called {callCount} times"
             );
+        }
+
+        [Fact]
+        public void GetStatesThrowsIncompleteBlockStatesException()
+        {
+            IStore store = _fx.Store;
+            Guid chainId = Guid.NewGuid();
+            var chain = new BlockChain<DumbAction>(
+                new NullPolicy<DumbAction>(),
+                store,
+                chainId
+            );
+            var privateKey = new PrivateKey();
+            var address = privateKey.PublicKey.ToAddress();
+
+            IImmutableDictionary<Address, object> GetDirty(
+                IEnumerable<ActionEvaluation<DumbAction>> evaluations) =>
+                evaluations.Select(
+                    a => a.OutputStates
+                ).Aggregate(
+                    ImmutableDictionary<Address, object>.Empty,
+                    (x, y) => x.SetItems(y.GetUpdatedStates())
+                );
+
+            // Build the store has incomplete states
+            Block<DumbAction> b = TestUtils.MineGenesis<DumbAction>();
+            chain.Blocks[b.Hash] = b;
+            store.IncreaseTxNonce(chainId.ToString(), b);
+            store.AppendIndex(chainId.ToString(), b.Hash);
+            IImmutableDictionary<Address, object> dirty =
+                GetDirty(b.Evaluate(DateTimeOffset.UtcNow, _ => null));
+            const int blocksCount = 5;
+            Address[] addresses = Enumerable.Repeat<object>(null, blocksCount)
+                .Select(_ => new PrivateKey().PublicKey.ToAddress())
+                .ToArray();
+            for (int i = 0; i < blocksCount; ++i)
+            {
+                Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
+                    store.GetTxNonce(chainId.ToString(), address),
+                    privateKey,
+                    new[] { new DumbAction(addresses[i], i.ToString()) }
+                );
+                b = TestUtils.MineNext(b, new Transaction<DumbAction>[] { tx });
+                dirty = GetDirty(
+                    b.Evaluate(DateTimeOffset.UtcNow, dirty.GetValueOrDefault)
+                );
+                Assert.NotEmpty(dirty);
+                chain.Blocks[b.Hash] = b;
+                store.StoreStateReference(
+                    chainId.ToString(),
+                    dirty.Keys.ToImmutableHashSet(),
+                    b
+                );
+                store.IncreaseTxNonce(chainId.ToString(), b);
+                store.AppendIndex(chainId.ToString(), b.Hash);
+            }
+
+            store.SetBlockStates(b.Hash, new AddressStateMap(dirty));
+
+            // As the store has the states for the tip (latest block),
+            // it shouldn't throw an exception.
+            Address lastAddress = addresses[blocksCount - 1];
+            AddressStateMap states = chain.GetStates(new[] { lastAddress });
+            Assert.NotEmpty(states);
+            Assert.Equal("4", states[lastAddress]);
+
+            // As the store lacks the states for blocks other than the tip,
+            // the following GetStates() calls should throw an exception.
+            foreach (Address addr in addresses.SkipLast(1))
+            {
+                Assert.Throws<IncompleteBlockStatesException>(() =>
+                    chain.GetStates(new[] { addr })
+                );
+            }
         }
 
         [Fact]