+++ CHANGES.md
@@ -18,11 +18,15 @@ To be released.
     `cancellationToken` option.  [[#287]]
  -  Added a `Peer` constructor omitting `appProtocolVersion` parameter
     to create a `Peer` whose version is unknown.
+ -  Added `IncompleteBlockStatesException` class.  [[#272], [#285]]
 
 ### Behavioral changes
 
  -  `BlockChain<T>.GetNonce()` became to count staged transactions too during
     nonce computation.  [[#270]]
+ -  `BlockChain<T>.GetStates()` method became to throw
+    `IncompleteBlockStatesException` if its `Store` lacks the states of a block
+    that a requested address lastly updated.  [[#272], [#285]]
  -  A message `Swarm` makes became to have multiple blocks within it, which
     means round trips on the network are now much reduced.  [[#273], [#276]]
  -  `Message.Block` has been replaced by `Message.Blocks` and the magic number
@@ -50,11 +54,13 @@ To be released.
 [#269]: https://github.com/planetarium/libplanet/pull/269
 [#270]: https://github.com/planetarium/libplanet/pull/270
 [#271]: https://github.com/planetarium/libplanet/pull/271
+[#272]: https://github.com/planetarium/libplanet/issues/272
 [#273]: https://github.com/planetarium/libplanet/issues/273
 [#275]: https://github.com/planetarium/libplanet/pull/275
 [#276]: https://github.com/planetarium/libplanet/pull/276
 [#277]: https://github.com/planetarium/libplanet/pull/277
 [#281]: https://github.com/planetarium/libplanet/pull/281
+[#285]: https://github.com/planetarium/libplanet/pull/285
 [#287]: https://github.com/planetarium/libplanet/pull/287